refactor(apollo): rename CheckboxCard component folder
