+++ client/apollo/react/src/indexLF.ts
@@ -41,7 +41,7 @@ export {
 } from "./CardMessage/CardMessageLF";
 export { ContentItemMono } from "./ContentItemMono/ContentItemMonoLF";
 export { Checkbox } from "./Form/Checkbox/Checkbox/CheckboxLf";
-export { CheckboxCard } from "./Form/Checkbox/checkboxCard/CheckboxCardLF";
+export { CheckboxCard } from "./Form/Checkbox/CheckboxCard/CheckboxCardLF";
 export { CheckboxText } from "./Form/Checkbox/checkboxText/CheckboxTextLF";
 export { DateInput } from "./Form/DateInput/DateInputLF";
 export { Heading, type HeadingLevel } from "./Heading/HeadingLF";
